Created a popup for component People
