+++ src/App.js
@@ -4,16 +4,13 @@ import React from "react";
 import "./index.scss";
 
 import { Header } from "./components";
-import { useSelector } from "react-redux";
 import { Route, Switch } from "react-router-dom";
 import { Films, People, Planets, Starships, Vehicles } from "./Pages";
 import { history } from "./redux/reducers";
 import { ConnectedRouter } from "connected-react-router";
 import { Species } from "./Pages/species/Species";
-import { Popup } from "./components/popup/Popup";
 
 const App = () => {
-	const isVisiblePopup = useSelector(state => state.app.isVisiblePopup);
 	return (
 		<>
 			<ConnectedRouter history={history}>
@@ -26,7 +23,6 @@ const App = () => {
 					<Route path="/vehicles" component={Vehicles} exact />
 					<Route path="/species" component={Species} exact />
 				</Switch>
-				{isVisiblePopup && <Popup />}
 			</ConnectedRouter>
 		</>
 	);
